# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.ResCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    format: int64
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
  controller.ResponseData:
    properties:
      code:
        $ref: '#/definitions/controller.ResCode'
      data: {}
      msg: {}
    type: object
  models.ParamCreatePost:
    properties:
      community_id:
        type: integer
      content:
        type: string
      title:
        type: string
    required:
    - community_id
    - content
    - title
    type: object
  models.ParamLoginEmail:
    properties:
      email:
        description: 直接利用 Gin 的 email 正则强校验
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  models.ParamLoginUsername:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.ParamSignUp:
    properties:
      password:
        type: string
      re_password:
        type: string
      username:
        type: string
    required:
    - password
    - re_password
    - username
    type: object
  models.ParamVoteData:
    properties:
      direction:
        description: 点赞(1) 点踩(-1) 取消投票(0)
        enum:
        - 1
        - 0
        - -1
        type: integer
      post_id:
        description: 帖子ID
        example: ""
        type: string
    required:
    - direction
    - post_id
    type: object
host: 127.0.0.1:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是基于 Gin 框架开发的社区后端 API 服务。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Echo 社区 API 文档
  version: "1.0"
paths:
  /community:
    get:
      consumes:
      - application/json
      description: 获取所有支持发帖的社区板块信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取所有社区板块列表
      tags:
      - 社区模块
  /community/{id}:
    get:
      consumes:
      - application/json
      description: 根据社区 ID 获取社区的详细介绍
      parameters:
      - description: 社区 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取指定社区详情
      tags:
      - 社区模块
  /login/email:
    post:
      consumes:
      - application/json
      description: 使用邮箱和密码登录
      parameters:
      - description: 邮箱登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamLoginEmail'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 邮箱登录
      tags:
      - 用户模块
  /login/username:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码登录
      parameters:
      - description: 登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamLoginUsername'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 用户名登录
      tags:
      - 用户模块
  /post:
    post:
      consumes:
      - application/json
      description: 登录用户可以在指定的社区板块内发布新帖子
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 发帖的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamCreatePost'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 创建新帖子
      tags:
      - 帖子模块
  /post/{id}:
    get:
      consumes:
      - application/json
      description: 根据帖子 ID 获取帖子的详细内容
      parameters:
      - description: 帖子 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取帖子详情
      tags:
      - 帖子模块
  /post/list:
    get:
      consumes:
      - application/json
      description: 支持按时间和分数排序。如果有 Token，会额外返回当前用户的点赞状态。
      parameters:
      - description: Bearer 用户令牌 (游客选填)
        in: header
        name: Authorization
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      - description: 排序规则 (time 或 score)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 获取帖子列表
      tags:
      - 帖子模块
  /post/vote:
    post:
      consumes:
      - application/json
      description: 登录用户对指定的帖子进行赞成(1)、反对(-1)或取消投票(0)操作
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 投票参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamVoteData'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 帖子投票
      tags:
      - 帖子模块
  /signup:
    post:
      consumes:
      - application/json
      description: 注册新用户，成功后直接返回 JWT Token 实现自动登录
      parameters:
      - description: 注册参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ParamSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: 用户注册
      tags:
      - 用户模块
swagger: "2.0"